backend/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
python-jose[cryptography]
python-multipart
pydantic[email]
python-dotenv
web3
ipfshttpclient
eth-account
py-solc-x